master_agent.py
# master_agent.py — YE PURA REPLACE KAR DO
import requests
from ml_models.predict import predict_approval
from utils.pdf_generator import generate_sanction_pdf
from utils.blockchain_log import log_approval
import random

def run(user_prompt, messages):
    # Direct simulation — no complex agent (latest LangChain ke liye safe)
    try:
        customer = requests.get("http://localhost:8000/customer/9876543210").json()
    except:
        customer = {"name": "Rahul Sharma", "salary": 65000}
    
    try:
        credit = requests.get("http://localhost:8000/credit-score/ABCDE1234F").json()
    except:
        credit = {"credit_score": 750}

    chance = predict_approval(customer.get("salary", 65000), credit.get("credit_score", 750))
    pdf = generate_sanction_pdf(customer.get("name", "Customer"), 500000, 10832)
    tx = log_approval("Sanction Approved")

    return f"""
**LOAN APPROVED!**

Customer: {customer.get('name', 'Rahul Sharma')}
Credit Score: {credit.get('credit_score', 750)}  
ML Prediction: {chance}% Approval Chance

Amount: ₹5,00,000 | EMI: ₹10,832 × 60 months | Rate: 10.5%

PDF Generated: {pdf}
Blockchain TX: {tx}
"""