requirements.txt
streamlit
langchain
langchain-groq
fastapi
uvicorn
reportlab
xgboost
pandas
web3
python-dotenv
python-multipart
pydantic